@@ -1,14 +1,11 @@
 name: google_ddns_update
-description: A sample command-line application.
-version: 1.1.0
+description: A command-line application to update Google Domains DDNS.
+version: 1.1.1
 author: Pedro Carvalho <inceptus.p@live.com>
 
 environment:
-  sdk: '>=2.5.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  args: ^1.6.0
-  http: ^0.12.2
-
-dev_dependencies:
-  pedantic: ^1.9.2
+  args: ^2.0.0
+  http: ^0.13.1
