@@ -1,61 +1,83 @@
 import 'dart:convert';
 import 'dart:io';
+
 import 'package:args/args.dart';
 import 'package:http/http.dart' as http;
 
-ArgResults argResults;
-
-void main(List<String> arguments) async {
-  final parser = ArgParser()
-    ..addOption('username', abbr: 'u', help: 'Provided Google DNS username', valueHelp: 'username')
-    ..addOption('password', abbr: 'p', help: 'Provided Google DNS password', valueHelp: 'password')
-    ..addOption('hostname',
-        abbr: 'n', help: 'Google DNS hostname', valueHelp: 'subdomain.domain.ttl')
+void main(List<String> args) async {
+  final argParser = ArgParser()
+    ..addOption(
+      'username',
+      abbr: 'u',
+      help: 'Provided Google Domains DDNS username',
+      valueHelp: 'username',
+    )
+    ..addOption(
+      'password',
+      abbr: 'p',
+      help: 'Provided Google Domains DDNS password',
+      valueHelp: 'password',
+    )
+    ..addOption(
+      'hostname',
+      abbr: 'n',
+      help: 'Google Domains DDNS hostname',
+      valueHelp: 'subdomain.domain.ttl',
+    )
     ..addFlag('help', abbr: 'h', help: 'This usage manual', negatable: false);
 
+  late ArgResults argResults;
+
   try {
-    argResults = parser.parse(arguments);
+    argResults = argParser.parse(args);
     if (argResults['help'] as bool) {
-      print('Usage:\n');
-      print(parser.usage + '\n');
+      print('\nUsage:\n');
+      print(argParser.usage + '\n');
       exit(0);
     }
   } catch (e) {
     print(e);
     print('\nUsage:\n');
-    print(parser.usage + '\n');
+    print(argParser.usage + '\n');
     exit(-1);
   }
 
-  var response = await http.get('https://api64.ipify.org?format=json');
-  Map<String, dynamic> publicIp;
+  var response = await http.get(Uri.parse('https://api64.ipify.org?format=json'));
   var lastIpFile = File('lastIP.txt');
-  String lastIp;
+  late Map<String, dynamic> publicIp;
+  late String lastIp;
 
   if (response.statusCode == 200) {
     publicIp = json.decode(response.body);
     if (lastIpFile.existsSync()) {
       lastIp = lastIpFile.readAsStringSync();
+    } else {
+      lastIp = '';
     }
-    lastIpFile.writeAsStringSync(publicIp['ip']);
     if (lastIp == publicIp['ip']) {
       print('IP haven\'t changed');
     } else {
       try {
-        var googleDns = await http.get('https://' +
-            argResults['username'] +
-            ':' +
-            argResults['password'] +
-            '@domains.google.com/nic/update?hostname=' +
-            argResults['hostname'] +
-            '&myip=' +
-            publicIp['ip']);
-        if (googleDns.statusCode == 200) {
-          print('Google says: ' + googleDns.body);
+        lastIpFile.writeAsStringSync(publicIp['ip']);
+        var googleDdnsResponse = await http.get(
+          Uri.parse(
+            'https://${argResults['username']}:${argResults['password']}@domains.google.com/nic/update?hostname=${argResults['hostname']}&myip=${publicIp['ip']}',
+          ),
+        );
+        if (googleDdnsResponse.statusCode == 200) {
+          print('Google says: ${googleDdnsResponse.body}');
+        } else {
+          print(
+            'Google domains HTTP error: ${googleDdnsResponse.statusCode} ${googleDdnsResponse.reasonPhrase}',
+          );
         }
       } catch (e) {
         print(e);
       }
     }
+  } else {
+    print(
+      'HTTP error when trying to get the IP Address: ${response.statusCode} ${response.reasonPhrase}',
+    );
   }
 }
